Add files via upload
Written Analysis of Results
1.	Overview of Project: Explain the purpose of this analysis.

The purpose of this exercise is to refactor the code to make the VBA script run faster and effectively build a stock table for analysis. 

2.	Results: Using images and examples of your code, compare the stock performance between 2017 and 2018, as well as the execution times of the original script and the refactored script.

The stock performance when comparing 2017 & 2018 reveals that stocks ENPH and RUN both showed a positive return overall while others had a decline in 2018.The stocks that performed poorly were DQ and JKS in 2018. 

The execution time was saved by 0.55 seconds by getting rid of the repetition of worksheet activation of All stock Analysis. It was moved from top to the bottom before part 4.

3.	Summary: In a summary statement, address the following questions.
a.	What are the advantages or disadvantages of refactoring code?
b.	How do these pros and cons apply to refactoring the original VBA script?

The advantage of refactoring a code is that it will help make the script run faster, reduce downtime, and make the script mor efficient. The disadvantage is that imprecise refactoring could introduce new bugs and errors into the code. In the case of larger teams working on refactoring, the coordination effort required could be surprisingly high.

The pro of refactoring is that it reduced the run time of the script. I was able to debug it and shift things around to see what till help reduce the overall time. The disadvantage was that I was running into errors and it was taking a lot of time to just save few seconds.